requirements.txt
APScheduler==3.10.4
aspose_words==24.5.0
Flask==3.0.3
google_api_python_client==2.122.0
protobuf==5.26.1
Requests==2.31.0
